requirements.txt
# AI/LLM 相关
anthropic>=0.40.0              # Claude API 客户端

# Web 框架
fastapi>=0.109.0               # 后端 API 框架
uvicorn>=0.27.0                # ASGI 服务器

# HTTP 请求
requests>=2.31.0               # 同步 HTTP 请求
aiohttp>=3.9.0                 # 异步 HTTP 请求

# 数据验证
pydantic>=2.5.0                # 数据模型验证

# 学术文献搜索
arxiv>=2.1.0                   # arXiv 论文搜索
semanticscholar>=0.7.0         # Semantic Scholar API

# 图片搜索
# ddgs 用于 DuckDuckGo 图片搜索（替代 serpapi）
ddgs>=6.0.0                    # DuckDuckGo 搜索库

# 文档处理
python-docx>=1.1.0             # Word 文档生成
docx2pdf>=0.1.8                # Word 转 PDF
PyMuPDF>=1.23.0                # PDF 处理（fitz）
Pillow>=10.0.0                 # 图片处理

# 图表生成
matplotlib>=3.7.0              # 数据可视化图表

# 工具库
markdown>=3.5.0                # Markdown 处理
beautifulsoup4>=4.12.0         # HTML 解析
python-dotenv>=1.0.0           # 环境变量加载
